+++ config/routes.rb
@@ -75,6 +75,7 @@ Rails.application.routes.draw do
     end
     resources :deliveries, only: %i[new destroy]
     resources :guild_invites, only: %i[new]
+    resources :settings, only: %i[index]
 
     root to: 'welcome#index'
   end