+++ frontend/packs/user_settings.jsx
@@ -1,0 +1,4 @@
+import UserSettings from "components_react/user_settings/user_settings";
+import WebpackerReact from "webpacker-react";
+
+WebpackerReact.setup({UserSettings});