+++ spec/controllers/settings_controller_spec.rb
@@ -1,0 +1,19 @@
+RSpec.describe SettingsController, type: :controller do
+  describe 'GET#index' do
+    it_behaves_like 'User Auth'
+
+    context 'for logged user' do
+      sign_in_user
+
+      it 'renders index template' do
+        get :index, params: { locale: 'ru' }
+
+        expect(response).to render_template :index
+      end
+    end
+
+    def do_request
+      get :index, params: { locale: 'en' }
+    end
+  end
+end