+++ spec/models/time_offset_spec.rb
@@ -1,0 +1,9 @@
+RSpec.describe TimeOffset, type: :model do
+  it { should belong_to :user }
+
+  it 'factory should be valid' do
+    time_offset = build :time_offset
+
+    expect(time_offset).to be_valid
+  end
+end