+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- settings page
+- TimeOffset model for user preferences about rendered time
+
 ## [0.7.4] - 2019-11-12
 ### Added
 - api endpoint recipes#search